blog_application/blog_app/likes.py
from blog_app.models import Like
from blog_app.serializers import LikeSerializer
from rest_framework import generics


class LikeList(generics.ListCreateAPIView):
    queryset = Like.objects.all()
    serializer_class = LikeSerializer


class LikeDetail(generics.RetrieveUpdateDestroyAPIView):
    queryset = Like.objects.all()
    serializer_class = LikeSerializer